# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpecForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specs:
    get:
      operationId: listSpecs
      tags: [specs]
      summary: List all generated specs
      responses:
        "200":
          description: List of specs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spec"
    post:
      operationId: createSpec
      tags: [specs]
      summary: Generate a new spec document using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecBody"
      responses:
        "201":
          description: Created spec
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spec"
  /specs/{id}:
    get:
      operationId: getSpec
      tags: [specs]
      summary: Get a specific spec document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spec document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spec"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteSpec
      tags: [specs]
      summary: Delete a spec document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /specs/{id}/stream:
    post:
      operationId: streamSpec
      tags: [specs]
      summary: Stream a spec document generation via SSE
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of generated spec content
          content:
            text/event-stream: {}
  /specs/{id}/chat:
    post:
      operationId: getOrCreateSpecChat
      tags: [specs]
      summary: Get or create a conversation for asking questions about a spec
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation for this spec
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /specs/share/{token}:
    get:
      operationId: getSpecByShareToken
      tags: [specs]
      summary: Get a publicly shared spec by its share token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared spec document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spec"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /specs/{id}/sync:
    post:
      operationId: syncSpec
      tags: [specs]
      summary: Manually trigger a re-generation of a spec from its original source
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncSpecResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /specs/{id}/webhook:
    get:
      operationId: getSpecWebhook
      tags: [specs]
      summary: Get or create webhook configuration for a spec
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Webhook config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookConfigResponse"
        "404":
          description: Not found
...
  /specs/{id}/share:
    post:
      operationId: shareSpec
      tags: [specs]
      summary: Generate or retrieve a public share link for a spec
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Share link details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareSpecResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TechDebtRisk:
      type: object
      properties:
        title:
          type: string
        severity:
          type: string
          enum: [high, medium, low]
        description:
          type: string
      required:
        - title
        - severity
        - description
    Spec:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        specType:
          type: string
          enum: [system_design, api_design, database_schema, feature_spec]
        inputType:
          type: string
          enum: [github_url, description]
        inputValue:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [pending, generating, completed, failed]
        complexityScore:
          type: integer
          nullable: true
        techDebtRisks:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/TechDebtRisk"
        complexitySummary:
          type: string
          nullable: true
        mermaidDiagram:
          type: string
          nullable: true
        conversationId:
          type: integer
          nullable: true
        shareToken:
          type: string
          nullable: true
        viewCount:
          type: integer
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - specType
        - inputType
        - inputValue
        - content
        - status
        - viewCount
        - createdAt
        - updatedAt
    ShareSpecResponse:
      type: object
      properties:
        shareToken:
          type: string
        shareUrl:
          type: string
        viewCount:
          type: integer
      required:
        - shareToken
        - shareUrl
        - viewCount
    CreateSpecBody:
      type: object
      properties:
        specType:
          type: string
          enum: [system_design, api_design, database_schema, feature_spec]
        inputType:
          type: string
          enum: [github_url, description]
        inputValue:
          type: string
        title:
          type: string
      required:
        - specType
        - inputType
        - inputValue
        - title
...
    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
...
    SyncSpecResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        lastSyncedAt:
          type: string
          nullable: true
      required:
        - id
        - status
    WebhookConfigResponse:
      type: object
      properties:
        webhookUrl:
          type: string
        secret:
          type: string
        instructions:
          type: string
      required:
        - webhookUrl
        - secret
        - instructions
...